rename `constant` to `constants`
